+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState.hs
@@ -23,7 +23,6 @@ module Cardano.Ledger.Shelley.LedgerState (
   InstantaneousRewards (..),
   LedgerState (..),
   PState (..),
-  PPUPPredFailure,
   RewardAccounts,
   RewardUpdate (..),
   RewardSnapShot (..),
@@ -158,5 +157,5 @@ import Cardano.Ledger.Shelley.LedgerState.PulsingReward
 import Cardano.Ledger.Shelley.LedgerState.Types
 import Cardano.Ledger.Shelley.PParams (pvCanFollow)
 import Cardano.Ledger.Shelley.RewardUpdate
-import Cardano.Ledger.Shelley.Rules.Ppup (PPUPPredFailure, ShelleyGovState (..))
+import Cardano.Ledger.Shelley.Rules.Ppup (ShelleyGovState (..))
 import Cardano.Ledger.Shelley.UTxO (consumed, produced)